Features two fields now for testing purposes + refactoring
The schema is now defined in a dedicated file located in a subdirectory
called "pplStock". There is a new field defined in the schema which needs
a parameter to generate its response.
